refactor: JdbcTemplate 내부 인터페이스 추출 및 메서드 로직 변경
