+++ src/main/java/web/dao/ResultSetExtractor.java
@@ -1,0 +1,9 @@
+package web.dao;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@FunctionalInterface
+interface ResultSetExtractor<T> {
+    T extractData(ResultSet rs) throws SQLException;
+}
