+++ src/main/java/web/dao/RowMapper.java
@@ -1,0 +1,9 @@
+package web.dao;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+@FunctionalInterface
+interface RowMapper<T> {
+    T map(ResultSet rs) throws SQLException;
+}
